+++ src/reader/_requests_utils.py
@@ -2,10 +2,13 @@
 Requests utilities. Contains no business logic.
 
 """
-from dataclasses import astuple
+from contextlib import contextmanager
+from contextlib import nullcontext
 from dataclasses import dataclass
 from dataclasses import field
 from typing import Any
+from typing import ContextManager
+from typing import Iterator
 from typing import Mapping
 from typing import Optional
 from typing import Protocol
@@ -42,16 +45,6 @@ class _ResponsePlugin(Protocol):
 
 
 @dataclass
-class SessionHooks:
-
-    request: Sequence[_RequestPlugin] = field(default_factory=list)
-    response: Sequence[_ResponsePlugin] = field(default_factory=list)
-
-    def copy(self: _T) -> _T:
-        return type(self)(*(list(v) for v in astuple(self)))
-
-
-@dataclass
 class SessionWrapper:
 
     """Minimal wrapper over requests.Sessions.
@@ -70,7 +63,8 @@ class SessionWrapper:
     """
 
     session: requests.Session = field(default_factory=requests.Session)
-    hooks: SessionHooks = field(default_factory=SessionHooks)
+    request_hooks: Sequence[_RequestPlugin] = field(default_factory=list)
+    response_hooks: Sequence[_ResponsePlugin] = field(default_factory=list)
 
     def get(
         self,
@@ -83,12 +77,12 @@ class SessionWrapper:
 
         request = requests.Request('GET', url, headers=headers)
 
-        for request_hook in self.hooks.request:
+        for request_hook in self.request_hooks:
             request = request_hook(self.session, request, **kwargs) or request
 
         response = self.session.send(self.session.prepare_request(request), **kwargs)
 
-        for response_hook in self.hooks.response:
+        for response_hook in self.response_hooks:
             new_request = response_hook(self.session, response, request, **kwargs)
             if new_request is None:
                 continue
@@ -115,6 +109,8 @@ class SessionWrapper:
 
 TimeoutType = Union[None, float, Tuple[float, float], Tuple[float, None]]
 
+DEFAULT_TIMEOUT = (3.05, 60)
+
 
 class TimeoutHTTPAdapter(requests.adapters.HTTPAdapter):
 
@@ -134,3 +130,50 @@ class TimeoutHTTPAdapter(requests.adapters.HTTPAdapter):
     def send(self, *args: Any, **kwargs: Any) -> Any:
         kwargs.setdefault('timeout', self.__timeout)
         return super().send(*args, **kwargs)
+
+
+@dataclass
+class SessionFactory:
+
+    """Manage the lifetime of a session."""
+
+    user_agent: Optional[str] = None
+    timeout: TimeoutType = DEFAULT_TIMEOUT
+    request_hooks: Sequence[_RequestPlugin] = field(default_factory=list)
+    response_hooks: Sequence[_ResponsePlugin] = field(default_factory=list)
+    session: Optional[SessionWrapper] = None
+
+    def make_session(self) -> SessionWrapper:
+        session = SessionWrapper(
+            request_hooks=list(self.request_hooks),
+            response_hooks=list(self.response_hooks),
+        )
+
+        timeout_adapter = TimeoutHTTPAdapter(self.timeout)
+        session.session.mount('https://', timeout_adapter)
+        session.session.mount('http://', timeout_adapter)
+
+        if self.user_agent:
+            session.session.headers['User-Agent'] = self.user_agent
+
+        return session
+
+    def transient(self) -> ContextManager[SessionWrapper]:
+        if self.session:
+            return nullcontext(self.session)
+        return self.make_session()
+
+    @contextmanager
+    def persistent(self) -> Iterator[SessionWrapper]:
+        # note: this is NOT threadsafe, but is reentrant
+
+        if self.session:  # pragma: no cover
+            yield self.session
+            return
+
+        with self.make_session() as session:
+            self.session = session
+            try:
+                yield session
+            finally:
+                self.session = None