+++ src/reader/plugins/ua_fallback.py
@@ -57,4 +57,4 @@ def _ua_fallback_response_hook(session, response, request, **kwargs):
 
 
 def init_reader(reader):
-    reader._parser.session_hooks.response.append(_ua_fallback_response_hook)
+    reader._parser.session_factory.response_hooks.append(_ua_fallback_response_hook)